chore: Move integration tests into tests dir
