+++ .github/workflows/rust.yml
@@ -19,4 +19,6 @@ jobs:
     - name: Build
       run: cargo build --verbose
     - name: Run tests
-      run: cargo test --verbose --features api-client,receipt-utility,ocsp
+      run: cargo test --verbose --features api-client,receipt-utility
+    - name: Run tests (Only ocsp)
+      run: cargo test --verbose --features ocsp
