+++ .gitignore
@@ -87,14 +87,7 @@ fabric.properties
 debug/
 target/
 
-# Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
-# More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
 Cargo.lock
-api/Cargo.lock
-core/Cargo.lock
-sdk/bindings/Cargo.lock
-sdk/example/Cargo.lock
-sdk/lib/Cargo.lock
 # These are backup files generated by rustfmt
 **/*.rs.bk
