+++ tests/resources/certs/SubscriptionKey_L256SYR32L.p8
