+++ tests/resources/mock_signed_data/missingX5CHeaderClaim
