+++ tests/resources/models/apiTooManyRequestsException.json
