+++ tests/resources/models/getAllSubscriptionStatusesResponse.json
