+++ tests/resources/models/signedConsumptionRequestNotification.json
