+++ tests/resources/models/signedExternalPurchaseTokenNotification.json
