+++ tests/resources/models/signedExternalPurchaseTokenSandboxNotification.json
