+++ tests/resources/models/signedSummaryNotification.json
