+++ tests/resources/models/transactionHistoryResponseWithMalformedAppAppleId.json
