+++ tests/resources/models/transactionHistoryResponseWithMalformedEnvironment.json
