+++ tests/resources/xcode/xcode-app-receipt-empty
