+++ tests/resources/xcode/xcode-app-receipt-legacy
