+++ tests/resources/xcode/xcode-app-receipt-with-transaction
