+++ tests/resources/xcode/xcode-signed-app-transaction
