+++ tests/resources/xcode/xcode-signed-renewal-info
