+++ tests/resources/xcode/xcode-signed-transaction
